Resolves conflicts with master's #6267 split of OnTreeCtrl_NavigatorSelectionChanged
into a handler plus UpdatePanelForItem helper, and drops Hide calls
for TextCtrl_Search/Button_Search/FlexGridSizer9 since master removed
those controls.

Also fixes the double-Thaw wx assert in RebuildTreeUI: the manual
Thaw call at the end ran on top of RebuildScopes destructor.

+++ src-iPad/Bridge/XLiPadInit.mm
@@ -86,7 +86,7 @@ static AnimatedImageData DecodeAnimatedImageIO(const uint8_t* data,
         // so the output matches xlImage's RGBA-interleaved layout directly.
         std::unique_ptr<uint8_t[]> rgba(new uint8_t[(size_t)w * h * 4]());
         CGContextRef ctx = CGBitmapContextCreate(rgba.get(), w, h, 8, w * 4, cs,
-                                                   kCGImageAlphaPremultipliedLast
+                                                   (uint32_t)kCGImageAlphaPremultipliedLast
                                                    | kCGBitmapByteOrder32Big);
         if (ctx) {
             CGContextSetBlendMode(ctx, kCGBlendModeCopy);
